Hw04 auth

+++ model/contacts.js
@@ -1,24 +1,54 @@
 const Contact = require('./schemas/contact')
 
-const getContacts = async () => {
-  return await Contact.find({})
+const getContacts = async (userId, query) => {
+  const {
+    limit = 5,
+    page = 1,
+    sortBy,
+    sortByDesc,
+    filter,
+    favorite = null,
+  } = query
+
+  const optionsSearch = { owner: userId }
+
+  if (favorite !== null) {
+    optionsSearch.favorite = favorite
+  }
+
+  const results = await Contact.paginate(optionsSearch, {
+    limit,
+    page,
+    select: filter ? filter.split('|').join(' ') : '',
+    sort: {
+      ...(sortBy ? { [`${sortBy}`]: 1 } : {}),
+      ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {}),
+    },
+  })
+
+  const { docs: contacts, totalDocs: total } = results
+
+  return { contacts, total, limit, page }
 }
 
-const getContactById = async contactId => {
-  return await Contact.findOne({ _id: contactId })
+const getContactById = async (userId, contactId) => {
+  return await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription -_id',
+  })
 }
 
 const createContact = async body => {
   return await Contact.create(body)
 }
 
-const removeContact = async contactId => {
-  return await Contact.findByIdAndRemove({ _id: contactId })
+const removeContact = async (userId, contactId) => {
+  return await Contact.findByIdAndRemove({ _id: contactId, owner: userId })
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId, contactId, body) => {
   return await Contact.findByIdAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true },
   )
